+++ NEWS.md
@@ -1,5 +1,25 @@
 # Version 0.1
 
+## Changes in 0.1.3
+
+- Now can choose or not to delete features already present when importing new
+  ones. Through parameter `deleteExistingFeatures =` in `import()`. Default is
+  `TRUE`.
+- Addition of functionalities to create, delete and list databases. 
+  See [#15](https://github.com/rformassspectrometry/RuSirius/issues/15)
+- Can now import feature with set adducts if known.
+  See [#14](https://github.com/rformassspectrometry/RuSirius/issues/14)
+- Small fixes for `show()` method of `Sirius` object.
+  See [#8](https://github.com/rformassspectrometry/RuSirius/issues/8)
+- Fix spectral Library matching and now can see a summary of full results using 
+  `summary()` and `results()` function and precising
+  `result.type = "spectralDbMatch"`.
+- Added results() for fragmentation tree. accessible using 
+  `result.type = "fragTree"`.
+  See [#18](https://github.com/rformassspectrometry/RuSirius/issues/18)
+- Added a DockerFile, but cannot seem to start Sirius properly. Work in Progress
+  See [#17](https://github.com/rformassspectrometry/RuSirius/issues/17)
+
 ## Changes in 0.1.2
 
 - Fix installation of Spectra dev version